[CCUBE-1134][KT] Move types into a file of its own in preparation for the schemas
